requirements.txt
pymarc==5.1.1
psycopg2==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
jsonpickle==3.0.2
saxonche==12.4.1.0
httpx==0.28.1
annif==1.3.1
unidecode==1.4.0